moved plugins to brain_bridge for helios_sim
