+++ helios_sim/src/brain_bridge/autonomy/components.rs
