+++ helios_sim/src/brain_bridge/autonomy/systems/mod.rs
@@ -3,9 +3,7 @@
 mod estimation;
 mod pipeline_tick;
 mod spawn;
-mod telemetry;
 
 pub use estimation::update_odom_frames;
 pub use pipeline_tick::run_pipeline_tick;
 pub use spawn::{spawn_autonomy_pipeline, spawn_odom_frames};
-pub use telemetry::{publish_autonomy_telemetry, publish_sensor_telemetry};
