--- helios_sim/src/plugins/autonomy/systems/telemetry.rs
@@ -1,9 +1,0 @@
-// helios_sim/src/simulation/plugins/autonomy/systems/telemetry.rs
-//
-// Telemetry stubs. Will be rewritten once the Zenoh observability layer lands (Step 9b+).
-
-/// No-op stub. Autonomy telemetry is deferred to the observability layer.
-pub fn publish_autonomy_telemetry() {}
-
-/// No-op stub. Sensor telemetry is deferred to the observability layer.
-pub fn publish_sensor_telemetry() {}
