+++ helios_sim/src/plugins/sensors/gps.rs
@@ -3,8 +3,8 @@ use nalgebra::Vector3;
 use rand_distr::{Distribution, Normal};
 use std::time::Duration;
 
+use crate::brain_bridge::autonomy::components::{AutonomyPipelineComponent, SensorPublishChannel};
 use crate::core::{app_state::SimulationSet, prng::SimulationRng, transforms::EnuVector};
-use crate::plugins::autonomy::components::{AutonomyPipelineComponent, SensorPublishChannel};
 use crate::prelude::*;
 
 use helios_core::data::envelope::SensorReading;
