+++ helios_sim/src/plugins/sensors/imu.rs
@@ -5,9 +5,9 @@ use nalgebra::Vector3;
 use rand_distr::{Distribution, Normal};
 use std::time::Duration;
 
+use crate::brain_bridge::autonomy::components::AutonomyPipelineComponent;
 use crate::core::transforms::EnuBodyPose;
 use crate::core::{app_state::SimulationSet, components::GroundTruthState, prng::SimulationRng};
-use crate::plugins::autonomy::components::AutonomyPipelineComponent;
 use crate::prelude::*;
 
 use helios_core::data::envelope::SensorReading;
