+++ helios_sim/src/plugins/sensors/magnetometer.rs
@@ -3,9 +3,9 @@ use nalgebra::Vector3;
 use rand_distr::{Distribution, Normal};
 use std::time::Duration;
 
+use crate::brain_bridge::autonomy::components::{AutonomyPipelineComponent, SensorPublishChannel};
 use crate::core::transforms::EnuBodyPose;
 use crate::core::{app_state::SimulationSet, prng::SimulationRng};
-use crate::plugins::autonomy::components::{AutonomyPipelineComponent, SensorPublishChannel};
 use crate::prelude::*;
 
 use helios_core::data::envelope::SensorReading;
